requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
python-dotenv==1.0.0
djangorestframework
djangorestframework-simplejwt

django-cors-headers

#swagger
drf-yasg
#pillow
Pillow==10.0.0
